Split DataSource in RemoteCommentDataSource and RemotePostDataSource.
