+++ app/src/main/java/ar/scacchipa/jsonplaceholderviewer/app/MyApplication.kt
@@ -7,7 +7,6 @@ class MyApplication: Application() {
     override fun onCreate() {
         super.onCreate()
         startKoin {
-            //androidContext(this@MyApplication)
             modules(appModule)
         }
     }
