+++ app/src/main/java/ar/scacchipa/jsonplaceholderviewer/data/IRemotePostSourceData.kt
@@ -1,0 +1,12 @@
+package ar.scacchipa.jsonplaceholderviewer.data
+
+import retrofit2.Response
+import retrofit2.http.GET
+
+interface IRemotePostSourceData {
+    @GET("posts")
+    suspend fun getPosts(): Response<List<Post>>
+}
+
+
+
