--- app/src/main/java/ar/scacchipa/jsonplaceholderviewer/data/ITypicodeSourceData.kt
@@ -1,27 +1,0 @@
-package ar.scacchipa.jsonplaceholderviewer.data
-
-import com.google.gson.GsonBuilder
-import retrofit2.Response
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.http.GET
-import retrofit2.http.Path
-
-interface ITypicodeSourceData {
-    @GET("posts")
-    suspend fun getPosts(): Response<List<Post>>
-
-    @GET("posts/{id}/comments")
-    suspend fun getComment(@Path("id") id: String): Response<List<Comment>>
-}
-
-fun provideRetrofit(): Retrofit {
-    return Retrofit.Builder()
-        .baseUrl("https://jsonplaceholder.typicode.com/")
-        .addConverterFactory(GsonConverterFactory.create(GsonBuilder().create()))
-        .build()
-}
-
-fun provideNetworkApi(retrofit: Retrofit): ITypicodeSourceData =
-    retrofit.create(ITypicodeSourceData::class.java)
-
